+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Breaking changes
 
 - (layout): removed Page, Container, Content and Sidebar (Sidebar, SidebarColumns, LogoContainer, SidebarPageTitle) components
+- (theme): inside toolbar the colors residing inside `icon` where moved to `item` since the colors applies also to labels and not only icons
 
 # v2.0.0-dev.7